updating ints

+++ ae/EnemyData.h
@@ -13,7 +13,7 @@ typedef struct {
 
 class EnemyData {
     public:
-        void Draw(Sphere_t *sObj, const uint_fast16_t MAX, uint_fast8_t *DFlag);
+        void Draw(Sphere_t *sObj, const uint_fast16_t MAX, uint_fast16_t *DFlag);
 };
 
 #endif
